Section 11 Arrays
Lesson 145 for each with maps and sets
-Maps returns the key and value pairs in the map through the three parameters and finally the map as a whole for example var.forEach(function(value, key, map){})

For sets you can use the same three parameters but the key and value return the same 'value' because sets dont work with keys. But forEach works the same for everything so the developers decided to return a duplicate value so at least the forEach working is consistent.