Section 11 Arrays
The reduce array method (most powerful array method) can be used to snowball / accumulate values in an array into one value and return that through a method. Also doable with a for loop but then you'll have to have multiple external variables when you have multiple loops. The method simply returns the value in the same variable. Reduce can be used to accumulate but also to search for the max or min value in an array. Always give up the initial value where the reduce should start (mostly the first value).